requirements.txt
streamlit
opencv-python
fer
numpy
Pillow
matplotlib
